# speedup range creation